+++ inst/graffle/classAndMethodAnalysis.R
@@ -10,7 +10,7 @@ Y <- lapply(
     content <- scan(file = filename, what = character(), sep = "\n", quiet = TRUE)
     hits <- stri_match_all(content, regex = '^setMethod\\("(.*?)"\\s*,\\s*"(.*?)".*')
     M <- do.call(rbind, lapply(hits, as.vector))
-    M[!is.na(M[,1]),]
+    M[!is.na(M[, 1]), ]
   })
 dt <- data.table(do.call(rbind, Y))
 dt <- dt[,2:3, with = FALSE]
@@ -21,5 +21,5 @@ colnames(dt) <- c("method", "class")
 dummy <- sapply(paste(dt[class == "cooccurrences"][["method"]], "\n", sep = ""), cat)
 
 # to crosstabulate methods / classes
-dt2 <- dcast(dt, class~method)
+dt2 <- dcast(dt, class ~ method)
 View(dt2)