+++ NEWS.md
@@ -13,9 +13,10 @@ polmineR 0.7.11.9001
 * `show()` and `size()`-method for `corpus` class.
 * `cpos()`, `count()`, `p_attributes()`,  `s_attributes()` `get_token_stream()` for `slice`-class, and for `subcorpus` via inheritance. 
 * The return value of the get_token_stream,regions-method was a data.table. The behavior is now in line with the other `get_token_stream() methods
-* new `coerce()`-method to turn `partition` object into `subcorpus` object
+* new `coerce()`-method to turn `partition` object into `subcorpus` object.
 * For all the S4 classes in the package multiple contains have been checked, and multiple contains have been removed.
 * A new experimental `remote_corpus`-class is the basis for accessing remote corpora.
+* The `subset()`-method for `corpus` and `subcorpus` objects has been reworked thoroughly, and now outperforms the `partition()` method. 
 
 polmineR 0.7.11
 ===============